moved source to lib/
